moved function from common to main, since it's specifically related to the generation of a kubeconfig
