+++ pkg/common/func.go
@@ -14,11 +14,8 @@ import (
 	"time"
 
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/containerd/containerd/log"
-
-	exec "github.com/alexellis/go-execute/pkg/v1"
 	"gopkg.in/yaml.v2"
 )
 
@@ -128,87 +125,6 @@ func NewInterLinkConfig() {
 
 		InterLinkConfigInst.set = true
 	}
-}
-
-func NewServiceAccount() error {
-
-	var sa string
-	var script string
-	path := InterLinkConfigInst.DataRootFolder + ".kube/"
-
-	err := os.MkdirAll(path, os.ModePerm)
-	if err != nil {
-		log.G(context.Background()).Error(err)
-		return err
-	}
-	f, err := os.Create(path + "getSAConfig.sh")
-	if err != nil {
-		log.G(context.Background()).Error(err)
-		return err
-	}
-
-	defer f.Close()
-
-	script = "#!" + InterLinkConfigInst.BashPath + "\n" +
-		"SERVICE_ACCOUNT_NAME=" + InterLinkConfigInst.ServiceAccount + "\n" +
-		"CONTEXT=$(kubectl config current-context)\n" +
-		"NAMESPACE=" + InterLinkConfigInst.Namespace + "\n" +
-		"NEW_CONTEXT=" + InterLinkConfigInst.Namespace + "\n" +
-		"KUBECONFIG_FILE=\"" + path + "kubeconfig-sa\"\n" +
-		"SECRET_NAME=$(kubectl get secret -l kubernetes.io/service-account.name=${SERVICE_ACCOUNT_NAME} --namespace ${NAMESPACE} --context ${CONTEXT} -o jsonpath='{.items[0].metadata.name}')\n" +
-		"TOKEN_DATA=$(kubectl get secret ${SECRET_NAME} --context ${CONTEXT} --namespace ${NAMESPACE} -o jsonpath='{.data.token}')\n" +
-		"TOKEN=$(echo ${TOKEN_DATA} | base64 -d)\n" +
-		"kubectl config view --raw > ${KUBECONFIG_FILE}.full.tmp\n" +
-		"kubectl --kubeconfig ${KUBECONFIG_FILE}.full.tmp config use-context ${CONTEXT}\n" +
-		"kubectl --kubeconfig ${KUBECONFIG_FILE}.full.tmp config view --flatten --minify > ${KUBECONFIG_FILE}.tmp\n" +
-		"kubectl config --kubeconfig ${KUBECONFIG_FILE}.tmp rename-context ${CONTEXT} ${NEW_CONTEXT}\n" +
-		"kubectl config --kubeconfig ${KUBECONFIG_FILE}.tmp set-credentials ${CONTEXT}-${NAMESPACE}-token-user --token ${TOKEN}\n" +
-		"kubectl config --kubeconfig ${KUBECONFIG_FILE}.tmp set-context ${NEW_CONTEXT} --user ${CONTEXT}-${NAMESPACE}-token-user\n" +
-		"kubectl config --kubeconfig ${KUBECONFIG_FILE}.tmp set-context ${NEW_CONTEXT} --namespace ${NAMESPACE}\n" +
-		"kubectl config --kubeconfig ${KUBECONFIG_FILE}.tmp view --flatten --minify > ${KUBECONFIG_FILE}\n" +
-		"rm ${KUBECONFIG_FILE}.full.tmp\n" +
-		"rm ${KUBECONFIG_FILE}.tmp"
-
-	_, err = f.Write([]byte(script))
-
-	if err != nil {
-		log.G(context.Background()).Error(err)
-		return err
-	}
-
-	//executing the script to actually retrieve a valid service account
-	cmd := []string{path + "getSAConfig.sh"}
-	shell := exec.ExecTask{
-		Command: "sh",
-		Args:    cmd,
-		Shell:   true,
-	}
-	execResult, _ := shell.Execute()
-	if execResult.Stderr != "" {
-		log.G(context.Background()).Error("Stderr: " + execResult.Stderr + "\nStdout: " + execResult.Stdout)
-		return errors.New(execResult.Stderr)
-	}
-
-	//checking if the config is valid
-	_, err = clientcmd.LoadFromFile(path + "kubeconfig-sa")
-	if err != nil {
-		log.G(context.Background()).Error(err)
-		return err
-	}
-
-	config, err := os.ReadFile(path + "kubeconfig-sa")
-	if err != nil {
-		log.G(context.Background()).Error(err)
-		return err
-	}
-
-	sa = string(config)
-	os.Remove(path + "getSAConfig.sh")
-	os.Remove(path + "kubeconfig-sa")
-
-	err = CreateClientsetFrom(context.Background(), sa)
-
-	return nil
 }
 
 func PingInterLink(ctx context.Context) (error, bool, int) {
